# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AuthCredentials {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
}

export interface ActivityImageItem {
  id: number;
  activityId: number;
  objectPath: string;
  altText: string;
  sortOrder: number;
}

export interface ActivityItem {
  id: number;
  title: string;
  description: string;
  stat: string;
  iconName: string;
  sortOrder: number;
  isActive: boolean;
  images: ActivityImageItem[];
}

export interface ActivityInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  stat?: string;
  iconName?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface ActivityUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  stat?: string;
  iconName?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface ActivityImageInput {
  objectPath: string;
  altText?: string;
  sortOrder?: number;
}

export interface ContentItem {
  id: number;
  key: string;
  value: string;
}

export interface ContentUpdate {
  value: string;
}

export interface UploadUrlInput {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResult {
  uploadURL: string;
  objectPath: string;
}