# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: activities
    description: Charity activities management
  - name: content
    description: Site content management
  - name: storage
    description: File storage
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Admin logout
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List all activities
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create a new activity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityItem"

  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get a single activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /activities/{id}/images:
    get:
      operationId: listActivityImages
      tags: [activities]
      summary: List images for an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityImageItem"
    post:
      operationId: addActivityImage
      tags: [activities]
      summary: Add image to activity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityImageInput"
      responses:
        "201":
          description: Created image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityImageItem"

  /activities/{activityId}/images/{imageId}:
    delete:
      operationId: deleteActivityImage
      tags: [activities]
      summary: Delete image from activity (admin)
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List all site content items
      responses:
        "200":
          description: Content items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"

  /content/{key}:
    patch:
      operationId: updateContent
      tags: [content]
      summary: Update a site content item (admin)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Presigned URL and object path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    AuthCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1

    AuthUser:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string

    ActivityImageItem:
      type: object
      required: [id, activityId, objectPath, altText, sortOrder]
      properties:
        id:
          type: integer
        activityId:
          type: integer
        objectPath:
          type: string
        altText:
          type: string
        sortOrder:
          type: integer

    ActivityItem:
      type: object
      required: [id, title, description, stat, iconName, sortOrder, isActive, images]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        stat:
          type: string
        iconName:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
        images:
          type: array
          items:
            $ref: "#/components/schemas/ActivityImageItem"

    ActivityInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        stat:
          type: string
        iconName:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean

    ActivityUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        stat:
          type: string
        iconName:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean

    ActivityImageInput:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string
        altText:
          type: string
        sortOrder:
          type: integer

    ContentItem:
      type: object
      required: [id, key, value]
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string

    ContentUpdate:
      type: object
      required: [value]
      properties:
        value:
          type: string

    UploadUrlInput:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string

    UploadUrlResult:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string