# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  stat?: string;
  iconName?: string;
  sortOrder?: number;
  isActive?: boolean;
}