# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityImageItem } from './activityImageItem';

export interface ActivityItem {
  id: number;
  title: string;
  description: string;
  stat: string;
  iconName: string;
  sortOrder: number;
  isActive: boolean;
  images: ActivityImageItem[];
}