# lib/api-zod/src/generated/types/activityUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  stat?: string;
  iconName?: string;
  sortOrder?: number;
  isActive?: boolean;
}