# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}